requirements.txt
asgiref==3.5.2
Django==4.0.6
django-ninja==0.19.0
psycopg2-binary==2.9.3
pydantic==1.9.1
sqlparse==0.4.2
typing_extensions==4.3.0
